requirements.txt
docker
numpy
pillow
pyyaml
requests